moved converter to new class
overhauled default options
